move polymer functions from auxiliary to services
