+++ Webapp/sources/auxiliary.py
@@ -8,7 +8,6 @@ from typing import Dict, List, Optional, Tuple, Type, Union
 
 from flask import abort, request
 from flask_login import current_user
-from psmiles import PolymerSmiles
 from sources import models
 from sources.extensions import db
 
@@ -40,161 +39,6 @@ def smiles_symbols(
     str
 ):  # restores the sharp plus and minus symbols from the string generated by marvinJS
     return compound.replace("minus", "-").replace("plus", "+").replace("sharp", "#")
-
-
-def extract_inside_brackets(
-    s: str,
-    start_index: int,
-) -> str:
-    """
-    Extracts contents of brackets to identify branching in polymer SMILES.
-    Handles nested branching.
-
-    Args:
-        s - whole SMILES string
-        start_index - index of first bracket
-
-    Returns:
-        The branch SMILES string and end index, or -1 if no branch found
-    """
-    balance = 0
-    inner_content = ""
-    for i in range(start_index, len(s)):
-        if s[i] == "(":
-            balance += 1
-        elif s[i] == ")":
-            balance -= 1
-
-        inner_content += s[i]  # Capture the character
-
-        if balance == 0:  # Found matching closing parenthesis
-            return inner_content, i
-
-    return inner_content, -1  # Return -1 if no matching parenthesis found
-
-
-def is_within_brackets(expression, target="{+n}"):
-    """
-    Checks if a target substring (like "{+n}") is within brackets.
-
-    Args:
-    - expression (str): String to search.
-    - target (str): The substring to check if it's within brackets.
-
-    Returns:
-    - bool: True if the target substring is within brackets, False otherwise.
-    """
-    # Find the position of the target in the expression
-    target_index = expression.find(target)
-
-    # Track the context of parentheses up to the target's position
-    balance = 0
-    for i in range(target_index):
-        if expression[i] == "(":
-            balance += 1
-        elif expression[i] == ")":
-            balance -= 1
-
-    # If we have open parentheses when reaching the target, it is within brackets
-    return balance > 0
-
-
-def get_last_bracketed_expression(expression):
-    """
-    Returns the last set of brackets, including its content, in a given expression.
-
-    Args:
-    - expression (str): The input string containing brackets.
-
-    Returns:
-    - str: The last complete set of brackets and its content, or an empty string if none found.
-    """
-    # Initialize variables to store positions of the last matching brackets
-    open_index = -1
-    close_index = -1
-    balance = 0
-
-    # Traverse the expression in reverse to find the last matching pair of brackets
-    for i in range(len(expression) - 1, -1, -1):
-        if expression[i] == ")":
-            if balance == 0:
-                close_index = i
-            balance += 1
-        elif expression[i] == "(":
-            balance -= 1
-            if balance == 0:
-                open_index = i
-                break
-
-    # If we have found a matching set, return the substring
-    if open_index != -1 and close_index != -1:
-        return expression[open_index + 1 : close_index]
-    return ""
-
-
-def reformat_smiles(smiles):
-    branch = smiles.split(")")[-1]
-    smiles = smiles[: smiles.rfind(branch)]
-    inner = get_last_bracketed_expression(smiles)
-    smiles = smiles.split(inner)[0] + branch + ")" + inner
-
-    return smiles
-
-
-def find_polymer_repeat_unit(
-    compound: str,
-) -> str:
-    """
-    Identify the repeat group within a polymer SMILES string.
-    Must be done before smiles is cleaned.
-    Uses {-} and {+n} to find repeat unit.
-    """
-    start_marker = compound.find("{-}")
-    end_marker = compound.find("{+n}")
-
-    # If both {-} and {+} are found
-    if start_marker != -1 and end_marker != -1:
-        # Reformat if end of SRU is treated as a branch. e.g. *C{-}(CC{+n}*)C where {+n} is within brackets
-        if is_within_brackets(compound, "{+n}"):
-            compound = reformat_smiles(compound)
-            start_marker = compound.find("{-}")
-            end_marker = compound.find("{+n}")
-
-        # Start is the character just before {-}, and end is the character just before {+}
-        result = compound[start_marker - 1] + compound[start_marker + 3 : end_marker]
-
-        # Check for branching: if the character after {+} is '('
-        if end_marker + 4 < len(compound) and compound[end_marker + 4] == "(":
-            branch, close_paren = extract_inside_brackets(compound, end_marker + 4)
-            if close_paren != -1:
-                result += branch
-
-            # check for more branching
-            while close_paren + 1 < len(compound) and compound[close_paren + 1] == "(":
-                branch, close_paren = extract_inside_brackets(compound, close_paren + 1)
-                if close_paren != -1:
-                    result += branch
-        return result
-    # if either {-} or {+} is not found
-    return ""
-
-
-def clean_polymer_smiles(
-    compound: str,
-) -> str:
-    """
-    Removes polymer symbols from the string generated by sketchers. last one removes dangling bonds
-    """
-    return compound.replace("{+n}", "").replace("{-}", "").replace("-", "")
-
-
-def canonicalise(smiles):
-    """
-    Canonicalise a polymer SMILES string. Not for use when endgroups are known (must have two '*'s)
-    """
-    ps = PolymerSmiles(smiles)
-    smiles = ps.canonicalize
-    return str(smiles).replace("[", "").replace("]", "")
 
 
 def get_data(field: str, request_data: Optional[Dict] = None) -> List[str]:
