+++ Webapp/sources/services/sustainability.py
@@ -2,8 +2,7 @@ from typing import Dict, List, Optional, Set
 
 import rdkit
 from flask import request
-from sources import models
-from sources.auxiliary import clean_polymer_smiles
+from sources import models, services
 from sources.extensions import db
 
 
@@ -129,7 +128,7 @@ def get_element_set(reaction_smiles_list: List):
     element_symbols = set()
     for component in reaction_smiles_list:
         if "*" in component:  # in polymer mode  # TODO: check
-            component = clean_polymer_smiles(component)
+            component = services.polymer_novel_compound.clean_polymer_smiles(component)
             component = component.replace(
                 "*", ""
             )  # not necessary to keep polymer * in smiles here
